Ubuntu series
Unify all the ubuntu maps, so it's easier to grok and maintain.
The supported bool on the seriesVersion defines if juju also supports
the ubuntu release or not. This will be later exposed, the idea of
this commit is to change the underlying code, but to not regress any
tests that we currently have.