Remove supported updates from distro info
The following commit removes the poly-fill from the distro info
csv file, when updating the Juju supported series. The basic idea
here is that Juju as a project needs to know what series it supports
and works with. New releases of ubuntu can throw up some new changes
that we might not of anticipated (netplan, systemd). So Juju will
need to have a better understanding and consideration around how
the Juju code works against new releases before allowing things to
be bootstrapped or deployed.

New series are still added to the supported series list for a couple
of reasons:
 1. We can use the series name for validation when deploying and
    bootstrapping
 2. We can identify those new os releases as we've added a new bool
    to highlight those if required.

It should be noted that all of this isn't relevant when the use of
`--force` is used in those scenarios. At that point the onus is on
the operator as they're bypassing the recommendations.