Add windows 2019 to os series
The following just adds windows 2019 to the windows os series.
This is required for the client tests which use windows 2019 server
for the github actions container.

Interestingly the windows side of this is broken as zero tests are
run when checking the commit/PR for validity. We should make a new
bot to check the windows tests.