DistroInfo source
The following encapsulates the distro info in a type that can be exposed
to juju directly. The series package defines functions at a package
level and it makes it hard to extract the data we need from with in
juju.

The code in this commit takes baby steps towards moving the code into a
type so we can access that from inside juju. It also provides backwards
compatibility to ensure that we don't break juju releases.

Interestingly, the series package is time sensitive (see
updateLocalSeriesVersions), where we don't pass a clock in to be able to
support fixing time to one spot.