Order supported series by version
The fact that we loose so much information when trying to gain
knownledge of a series makes it hard to make rational descisions about
which series to pick.

Rather that re-write this god-awful package, we can order the series and
then use that to at least make the results predictable.