Remove the default supported lts function
The function has been moved to `juju/juju` directly as this should be
locked stepped with juju, this isn't just a default supported lts, but
what juju uses as the default supported series.

It was a mistake to think that they were independent and this change
means that the os package can move without fearing that it might cause
juju to fail when deploying code.