Get OS from series using a hint
The following function helps provide a hint to help when normalising the
OS from a series. For OSes like centos that just have a number after it,
then the series of centos7 is fine, but fails when someone wants to just
pass 7. Unless we know the potential os as a hint (centos) then we can
then correctly normalise the value.